add redis cache and fix logout bugs

+++ fir_ser/api/views/logout.py
@@ -1,6 +1,7 @@
 from rest_framework.views import APIView
 from api.utils.auth import ExpiringTokenAuthentication
 from api.models import Token
+from django.core.cache import cache
 from rest_framework.response import Response
 
 from api.utils.permission import LoginUserPermission
@@ -12,5 +13,7 @@ class LogoutView(APIView):
 
     def delete(self, request):
         user = request.user.pk
-        Token.objects.filter(user=user).delete()
+        auth_token = request.auth
+        cache.delete(auth_token)
+        Token.objects.filter(user=user,access_token=auth_token).delete()
         return Response({"code": 1000})
